pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-labeling-server"
version = "0.1.0"
description = "Fast, configurable web server for image labeling with keyboard shortcuts"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Mitch", email = "mitch@example.com"}
]
keywords = ["labeling", "annotation", "machine-learning", "image-classification"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "pydantic>=2.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
ml = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "scikit-learn>=1.3.0",
    "timm>=0.9.0",
]

[project.scripts]
image-labeling-server = "smart_label.__main__:main"

[project.urls]
Homepage = "https://github.com/mitchins/image-labelling-server"
Repository = "https://github.com/mitchins/image-labelling-server"
Issues = "https://github.com/mitchins/image-labelling-server/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["smart_label*"]
exclude = ["tests*", "*.egg-info"]

[tool.pytest.ini_options]
testpaths = ["smart_label/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "N", "UP"]
